Adding dcr service extension implementation
