+++ financial-services-accelerator/accelerators/fs-is/repository/resources/wso2is-7.1.0-deployment.toml
@@ -280,7 +280,6 @@ grant_handler = "org.wso2.financial.services.accelerator.identity.extensions.gra
 [oauth.dcr]
 enable_fapi_enforcement=true
 ssa_jkws= "https://keystore.openbankingtest.org.uk/0015800001HQQrZAAX/0015800001HQQrZAAX.jwks"
-additional_attribute_filter= "org.wso2.financial.services.accelerator.identity.extensions.dcr.attribute.filter.FSAdditionalAttributeFilter"
 
 [oauth.oidc]
 id_token.signature_algorithm="PS256"
@@ -326,7 +325,7 @@ data_source.name = "WSO2FS_DB"
 connection_verification_timeout=1
 
 [financial_services.app_registration]
-applicationupdater = "org.wso2.financial.services.accelerator.identity.extensions.dcr.application.listener.ApplicationUpdaterImpl"
+applicationupdater = "listener.application.org.wso2.financial.services.accelerator.identity.extensions.client.registration.ApplicationUpdaterImpl"
 
 [[financial_services.app_registration.dcr.params]]
 name = "SoftwareId"
@@ -387,48 +386,48 @@ include_in_response = true
 
 [[financial_services.app_registration.dcr.validators.validator]]
 name = "RequiredParamsValidator"
-class = "org.wso2.financial.services.accelerator.identity.extensions.dcr.validators.RequiredParamsValidator"
+class = "validators.org.wso2.financial.services.accelerator.identity.extensions.client.registration.RequiredParamsValidator"
 enable = true
 priority = 1
 [[financial_services.app_registration.dcr.validators.validator]]
 name = "IssuerValidator"
-class = "org.wso2.financial.services.accelerator.identity.extensions.dcr.validators.IssuerValidator"
+class = "validators.org.wso2.financial.services.accelerator.identity.extensions.client.registration.IssuerValidator"
 enable = true
 priority = 2
 [[financial_services.app_registration.dcr.validators.validator]]
 name = "RedirectUriFormatValidator"
-class = "org.wso2.financial.services.accelerator.identity.extensions.dcr.validators.RedirectUriFormatValidator"
+class = "validators.org.wso2.financial.services.accelerator.identity.extensions.client.registration.RedirectUriFormatValidator"
 enable = true
 priority = 3
 [[financial_services.app_registration.dcr.validators.validator]]
 name = "RedirectUriMatchValidator"
-class = "org.wso2.financial.services.accelerator.identity.extensions.dcr.validators.RedirectUriMatchValidator"
+class = "validators.org.wso2.financial.services.accelerator.identity.extensions.client.registration.RedirectUriMatchValidator"
 enable = true
 priority = 4
 [[financial_services.app_registration.dcr.validators.validator]]
 name = "UriHostnameValidator"
-class = "org.wso2.financial.services.accelerator.identity.extensions.dcr.validators.UriHostnameValidator"
+class = "validators.org.wso2.financial.services.accelerator.identity.extensions.client.registration.UriHostnameValidator"
 enable = true
 priority = 5
 [[financial_services.app_registration.dcr.validators.validator]]
 name = "SSAIssuerValidator"
-class = "org.wso2.financial.services.accelerator.identity.extensions.dcr.validators.SSAIssuerValidator"
+class = "validators.org.wso2.financial.services.accelerator.identity.extensions.client.registration.SSAIssuerValidator"
 enable = true
 priority = 6
 allowed_values = ["OpenBanking Ltd"]
 [[financial_services.app_registration.dcr.validators.validator]]
 name = "RequestJTIValidator"
-class = "org.wso2.financial.services.accelerator.identity.extensions.dcr.validators.RequestJTIValidator"
+class = "validators.org.wso2.financial.services.accelerator.identity.extensions.client.registration.RequestJTIValidator"
 enable = true
 priority = 7
 [[financial_services.app_registration.dcr.validators.validator]]
 name = "SSAJTIValidator"
-class = "org.wso2.financial.services.accelerator.identity.extensions.dcr.validators.SSAJTIValidator"
+class = "validators.org.wso2.financial.services.accelerator.identity.extensions.client.registration.SSAJTIValidator"
 enable = true
 priority = 8
 [[financial_services.app_registration.dcr.validators.validator]]
 name = "TokenEndpointAuthSigningAlgValidator"
-class = "org.wso2.financial.services.accelerator.identity.extensions.dcr.validators.TokenEndpointAuthSigningAlgValidator"
+class = "validators.org.wso2.financial.services.accelerator.identity.extensions.client.registration.TokenEndpointAuthSigningAlgValidator"
 enable = true
 priority = 9
