+++ financial-services-accelerator/components/org.wso2.financial.services.accelerator.identity.extensions/src/main/java/org/wso2/financial/services/accelerator/identity/extensions/internal/IdentityExtensionsServiceComponent.java
@@ -39,14 +39,12 @@ import org.wso2.carbon.identity.oauth2.OAuth2Service;
 import org.wso2.carbon.identity.openidconnect.ClaimProvider;
 import org.wso2.carbon.identity.openidconnect.RequestObjectService;
 import org.wso2.carbon.user.core.service.RealmService;
-import org.wso2.financial.services.accelerator.common.config.FinancialServicesConfigParser;
 import org.wso2.financial.services.accelerator.common.config.FinancialServicesConfigurationService;
-import org.wso2.financial.services.accelerator.common.constant.FinancialServicesConstants;
-import org.wso2.financial.services.accelerator.common.util.FinancialServicesUtils;
 import org.wso2.financial.services.accelerator.consent.mgt.service.ConsentCoreService;
 import org.wso2.financial.services.accelerator.identity.extensions.claims.FSClaimProvider;
 import org.wso2.financial.services.accelerator.identity.extensions.claims.RoleClaimProviderImpl;
-import org.wso2.financial.services.accelerator.identity.extensions.dcr.application.listener.FSApplicationManagementListener;
+import org.wso2.financial.services.accelerator.identity.extensions.client.registration.application.listener.FSApplicationManagementListener;
+import org.wso2.financial.services.accelerator.identity.extensions.client.registration.dcr.attribute.filter.FSAdditionalAttributeFilter;
 
 /**
  * Identity common data holder.
@@ -69,10 +67,7 @@ public class IdentityExtensionsServiceComponent {
         bundleContext.registerService(ClaimProvider.class.getName(), new RoleClaimProviderImpl(), null);
 
         if (Boolean.parseBoolean(IdentityUtil.getProperty("OAuth.DCRM.EnableFAPIEnforcement"))) {
-            String filterConfig = (String) FinancialServicesConfigParser.getInstance().getConfiguration()
-                    .get(FinancialServicesConstants.DCR_ADDITIONAL_ATTRIBUTE_FILTER);
-            AdditionalAttributeFilter attributeFilter = FinancialServicesUtils
-                    .getClassInstanceFromFQN(filterConfig, AdditionalAttributeFilter.class);
+            AdditionalAttributeFilter attributeFilter = new FSAdditionalAttributeFilter();
             bundleContext.registerService(AdditionalAttributeFilter.class.getName(), attributeFilter, null);
         }
